move zone manager ui to new file
